project/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
celery==5.1.2
redis==3.5.3
requests==2.27.1
drf-yasg==1.20.0
python-dotenv